Handle warnings (#6)
* Set version to 7.0.0

* Remove NDEBUG and BOOST_UBLAS_NDEBUG from source files to CMakePresets.json

* TODO: Figure out why checks are conditioned on THETA_RES being set

* Remove dead code

* Only set Theta_res if THETA_RES is defined

* Delete unused variables

* Remove -Wno-mismatched-tags

* Fixed some straightforward sign-compare warnings

* Handle some of the unusued parameter warnings

Changed from named to unnamed parameter when reason for not using
is obvious.
Add FIXME tag when reason is not obvious

* Add FIXME to set-but-unused variables

* Dont use default move constructor

The default move constructor is implicitly deleted because there is a
unique_ptr, which has a deleted move constructor

* Delete FoundEntry copy constructor so implicit is used

Defining an implicit copy asignment operator when there is an explicit
copy constructor is deprecated. Since the explicit copy constructor is
just doing what the implicit will do, we delete it.

* Fix mismatched-tags warnings

* Delete Entry copy assignment operator

Implicit copy constructors are deprecated when copy assignment
operator is defined. The Entry copy assignment operator can be
replaced with an implicit, so we just delete it

* Collapse all #if 1 blocks

Delete all code in else part
Delete all #if 1 / #else / #endif statements

* Remove all #if 0 blocks from header files

* Delete #if 0 blocks from source files

* Delete #else part of #if 1