Refactor surface model
Split surface model into three parts, interface, model and component.
The interface part defines the surface model API.
The model part implements various the surface models.
The component part connects the models to Daisy's object model.

The refactoring leads to an issue when using TREELOG_MODEL, since it is used
in a surface model but assumes it is called from a surface component. We should
figureout if TREELOG_MODEL provides any value, and if so figure out how to get it
back.

For now we keep the Surface class as a super class for surface components,
but at some point we want to switch so other models that depend on Surface,
depends on a Surface model and not on a Surface component.