Try to fix MacOS dylib load bug
On some macs we get relative dylib paths for libcxsparse and
libsuitesparseconfig, but absolute dylib paths for libboost*

The relative paths are @rpath/<lib-name>, so we now convert all
absolute paths to this format instead of to @rpath/lib/<lib-name>
and then add @executable_path/lib (or @executable_path/bin/lib) as
an rpath.

We also set CMAKE_BUILD_RPATH_USE_ORIGIN=OFF to force absolute paths
At some point it would be nicer to set this to on so we dont have to
figure out the absolute path, but support for it seems limited.