N2 o (#30)
* add plots of buildable function models in ref manual

* fix p96 functions and add equation to ref manual

* denit n2o runs now

* Include missing dlls in win installer + update README

libsuitesparseconfig.dll and libgomp-1.dll needs to be included in
windows installer.

Added instructions for getting test framework installed on windows

* Rename project to daisy-N2O + fix hardcoded win reg key name

We need a workflow where we can test new components/models. This
should probably be a plugin based architecture, but for now we just
use a different project name.

On windows the project name is used as part of the registry key that
stores the location of daisy lib/ and sample/ directories. The key
used in path.C was previously partly hardcoded and thus failed when
the project name is changed.

* (SVAT_SSOC::calculate_conductances): Fix argument order (reported by Delhez Laura <Laura.Delhez@uliege.be>).

* add N2O denit to field nitrogen log

* New baseline files with new denit columns

---------

Co-authored-by: Per Abrahamsen <pa@plen.ku.dk>
Co-authored-by: Silas Ørting <silas.orting@pm.me>
Co-authored-by: Silas <silas@di.ku.dk>